Move SecurityService to model and update references
